WFLY-1: Installed browserrouter, refactored the structure.
