+++ src/components/App.js
@@ -1,15 +1,20 @@
 import React from 'react';
-import Landing from './Landing';
-import ClosetrFeature from './ClosetrFeature';
+import { BrowserRouter, Route } from 'react-router-dom';
+import Home from './home/Home';
+import Team from './Team';
+import ClosetrDetail from './work/ClosetrDetail';
 import "../styles/styles.scss";
 
 class App extends React.Component {
   render() {
     return (
-      <div>
-        <Landing/>
-        <ClosetrFeature/>
-      </div>
+      <BrowserRouter>
+        <div>
+          <Route path="/" exact component={Home}/>
+          <Route path="/about" exact component={Team}/>
+          <Route path="/closetr" exact component={ClosetrDetail}/>
+        </div>
+      </BrowserRouter>
     );
   }
 }
