+++ src/components/Team.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+class Team extends React.Component {
+  render() {
+    return (
+      <div>team</div>
+    );
+  }
+}
+
+export default Team;
