+++ src/components/home/Landing.js
@@ -1,7 +1,7 @@
 import React from 'react';
-import logo from '../assets/waffly-logo.png';
-import landingArt from '../assets/waffly-landing-art.png';
-import arrowDown from '../assets/arrow-down.svg';
+import logo from '../../assets/waffly-logo.png';
+import landingArt from '../../assets/waffly-landing-art.png';
+import arrowDown from '../../assets/arrow-down.svg';
 
 class Landing extends React.Component {
   render() {
