+++ src/components/work/ClosetrDetail.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+class ClosetrDetail extends React.Component {
+  render() {
+    return (
+      <div>closetr detail</div>
+    );
+  }
+}
+
+export default ClosetrDetail;
